+++ .idea/misc.xml
@@ -6,10 +6,14 @@
         <entry key="..\:/Users/Rajat/Downloads/Face-Attendance/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.3828125" />
         <entry key="..\:/Users/Rajat/Downloads/Face-Attendance/app/src/main/res/drawable/ic_launcher_background.xml" value="0.3828125" />
         <entry key="..\:/Users/Rajat/Downloads/Face-Attendance/app/src/main/res/layout/activity_main.xml" value="0.2957427536231884" />
+        <entry key="..\:/Users/grove/AndroidStudioProjects/Face-Attendance/app/src/main/res/layout/activity_main.xml" value="0.264" />
+        <entry key="..\:/Users/grove/AndroidStudioProjects/Face-Attendance/app/src/main/res/layout/activity_student_menu.xml" value="0.13903985507246377" />
+        <entry key="..\:/Users/grove/AndroidStudioProjects/Face-Attendance/app/src/main/res/layout/activity_upload_image.xml" value="0.14583333333333334" />
+        <entry key="..\:/Users/grove/AndroidStudioProjects/Face-Attendance/app/src/main/res/layout/activity_uploadsection.xml" value="0.14583333333333334" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
