+++ app/src/main/java/com/softwareengineering/faceattendance/MainActivity.java
@@ -36,12 +36,11 @@ public class MainActivity extends AppCompatActivity {
 
     ImageView img;
     Button upload, takephoto;
-
+    EditText roll, name;
     Uri filepath;
 
     Bitmap bitmap;
-    EditText roll;
-
+    Button login, signup;
 
 
     @Override
@@ -49,105 +48,26 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        TextView textView;
+        login = (Button) findViewById(R.id.login_Id);
+        signup = (Button) findViewById(R.id.sign_up_id);
 
-        img = (ImageView)findViewById(R.id.imageview_id);
-        upload = (Button) findViewById(R.id.upload_id);
-        takephoto = (Button) findViewById(R.id.gallery_id);
-        roll = (EditText) findViewById(R.id.rollno_id);
-
-
-        takephoto.setOnClickListener(new View.OnClickListener() {
+        signup.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Dexter.withActivity(MainActivity.this)
-                      .withPermission(Manifest.permission.READ_EXTERNAL_STORAGE)
-                      .withListener(new PermissionListener() {
-                          @Override
-                          public void onPermissionGranted(PermissionGrantedResponse permissionGrantedResponse) {
-                              Intent intent = new Intent(Intent.ACTION_PICK);
-                              intent.setType("image/*");
-                              startActivityForResult(Intent.createChooser(intent, "Select Image"), 1);
-                          }
+                Intent intent = new Intent(MainActivity.this, UploadImage.class);
+                startActivity(intent);
+            }
+        });
 
-                          @Override
-                          public void onPermissionDenied(PermissionDeniedResponse permissionDeniedResponse) {
-
-                          }
-
-                          @Override
-                          public void onPermissionRationaleShouldBeShown(PermissionRequest permissionRequest, PermissionToken permissionToken) {
-                              permissionToken.continuePermissionRequest();
-                          }
-                      }).check();
-
+        login.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, StudentMenu.class);
+                startActivity(intent);
             }
         });
 
 
-        upload.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                   uploadtofirebase();
-
-            }
-        });
-
-    }
-
-
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        if(requestCode == 1 && resultCode == RESULT_OK)
-        {
-           filepath = data.getData();
-
-           try {
-               InputStream inputStream = getContentResolver().openInputStream(filepath);
-               bitmap = BitmapFactory.decodeStream(inputStream);
-
-               img.setImageBitmap(bitmap);
-           }
-
-           catch (Exception ex){
-
-           }
-        }
-
-        super.onActivityResult(requestCode, resultCode, data);
-    }
-
-
-
-    private void uploadtofirebase()
-    {
-        final ProgressDialog dialog = new ProgressDialog(this);
-        dialog.setTitle("File Uploader");
-        dialog.show();
-
-        FirebaseStorage storage = FirebaseStorage.getInstance();
-        StorageReference uploader = storage.getReference("faces/" + roll.getText().toString());
-        uploader.putFile(filepath)
-                .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>()
-                {
-                    @Override
-                    public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-
-                        dialog.dismiss();
-
-                        Toast.makeText(getApplicationContext(), "File Uploaded", Toast.LENGTH_LONG).show();
-                    }
-                })
-
-                .addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
-                    @Override
-                    public void onProgress(@NonNull UploadTask.TaskSnapshot taskSnapshot)
-                    {
-                      float percent = (100 * taskSnapshot.getBytesTransferred())/taskSnapshot.getTotalByteCount();
-                      dialog.setMessage("Uploaded :" +(int)percent+ "%");
-                    }
-                });
 
     }
 }
