+++ app/src/main/java/com/softwareengineering/faceattendance/StudentMenu.java
@@ -1,0 +1,16 @@
+package com.softwareengineering.faceattendance;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class StudentMenu extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_student_menu);
+
+
+    }
+}
